mmr-testing/mediasource-seekable-forMMR.html

<!DOCTYPE html>
<!-- Copyright © 2016 Chromium authors and World Wide Web Consortium, (Massachusetts Institute of Technology, ERCIM, Keio University, Beihang). -->
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="mediasource-util.js"></script>
<script>
    mediasource_test(function(test, mediaElement, mediaSource)
    {
        mediaElement.addEventListener('error', test.unreached_func("Unexpected event 'error'"));
        mediaElement.addEventListener('ended', test.step_func_done(function () {}));

        var sourceBuffer = mediaSource.addSourceBuffer(MediaSourceUtil.AUDIO_ONLY_TYPE);

        assertSeekableEquals(mediaElement, '{ }', 'mediaElement.seekable');
        test.done();
    }, 'Get seekable time ranges when the sourcebuffer is empty.');

    mediasource_testafterdataloaded(function(test, mediaElement, mediaSource, segmentInfo, sourceBuffer, mediaData)
    {
        var initSegment = MediaSourceUtil.extractSegmentData(mediaData, segmentInfo.init);
        test.expectEvent(mediaElement, 'durationchange', 'mediaElement got duration');
        sourceBuffer.appendBuffer(initSegment);
        test.waitForExpectedEvents(function()
        {
            assert_equals(mediaElement.duration, segmentInfo.duration);
            assertSeekableEquals(mediaElement, '{ [0.000, ' + segmentInfo.duration.toFixed(3) + ') }', 'mediaElement.seekable');
            test.done();
        });
    }, 'Get seekable time ranges after init segment received.');

</script>
